Improvements...
Added a m_cycleTime variable that is calculated each main-loop cycle, it
represents the time it takes for the bot to execute one cycle in
milliseconds.

Also, added requests for command & tag messages. Now the bot asks the
twitch server for membership, command and tag messages, waits for the
correct responses and exits the program with an errorcode if that fails.